rename: 홈페이지 컴포넌트 containers로 이동
Related to: #32
