+++ src/containers/home/ActivityCardSlider.tsx
