+++ src/containers/home/ContestBanner.tsx
