+++ src/containers/home/ContestCardSlider.tsx
